Fruit Ninja
Created the class “Fruit.” There are now two watermelon images, the
sliced version and the not sliced version. A boolean determines if the
watermelon is sliced or not and therefore which image is showing. As of
now, only a single watermelon falls at once, and we will work on adding
more fruits later. The watermelon falls from a random place on top of
the screen.